// Live/frontend/src/client/models/ItemCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ItemCreate = {
    id: string;
    surname :string;
    first_name: string;
    last_name: string;
    address: string;
    house_no: string;
    street: string;
    landmark: string;
    native_place : string;
    post_office: string;
    area: string;
    district: string;
    sub_district: string;
    state: string;
    email: string;
    mobile_country_code: string;
    mobile : string;
    pin_code : BigInteger;
    note: string; 
    description?: (string | null);    
};